+++ gpaw/__init__.py
@@ -9,7 +9,7 @@ from pathlib import Path
 from typing import List, Union, Any, TYPE_CHECKING
 
 __version__ = '24.5.0'
-__ase_version_required__ = '3.22.1'
+__ase_version_required__ = '3.23.0'
 
 __all__ = ['GPAW',
            'Mixer', 'MixerSum', 'MixerDif', 'MixerSum2',